feat: split core to be able to use mipsum in others platforms, like a cli interface, an web API and so on...
