+++ src/index.js
@@ -1,4 +1,5 @@
-import { MussumIpsum, mIpsum } from './lib/mipsum'
+import MussumIpsum from './lib/mipsum-core'
+import { mIpsum } from './lib/mipsum'
 
 export {
   MussumIpsum,
